don't create an SSL context if not needed
Specifically, don't create one (and don't use it) if neither
VAULT_CACERT or VAULT_CAPATH are defined.

This narrows the Python >= 2.7.9 requirement. Now earlier versions work
if there is no SSL context, because we're not calling
ssl.create_default_context(), and we're not passing a context attribute
to urllib2.urlopen(), which is also new as of 2.7.9:
https://docs.python.org/2/library/urllib2.html#urllib2.urlopen