variable lookup works with ansible 1.9.x and 2.x
This was just getting variables from 'inject', which is set for 1.9.x
only. ansible 2.x passes variables via 'variables'. Choose one or the
other with '(variables or inject)'. These are mutually exclusive, so we
can just use one or the other and don't need the union.

Also use the get() method of a dict to more cleanly handle both None
situations (that would otherwise lead to TypeError) and the key not
being present (that would otherwise lead to KeyError):
https://docs.python.org/2/library/stdtypes.html#dict.get